refactor(desktop): extract Composer cluster (Phase 4/5)
- Composer/ModelCombobox.tsx
- Composer/ThinkingValueSelect.tsx
- Composer/index.tsx — Composer with its ~50-prop ComposerProps
  surface preserved
